Relocate config
