+++ lib/lit-protocol-modal/chainConfig.js
@@ -1,5 +1,5 @@
-import { ethereumTypesConfig } from "./ethereum/ethereumTypesConfig";
-import { solanaTypesConfig } from "./solana/solanaTypesConfig";
+import { ethereumTypesConfig } from "./chainComponents/ethereum/ethereumTypesConfig";
+import { solanaTypesConfig } from "./chainComponents/solana/solanaTypesConfig";
 
 import { utils } from 'ethers';
